[Fix] directory structre & norm
