+++ srcs/parse/lexer.c
@@ -6,7 +6,7 @@
 /*   By: jaekpark <jaekpark@student.42seoul.fr      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/04 18:17:44 by jaekpark          #+#    #+#             */
-/*   Updated: 2021/07/26 20:43:56 by parkjaekw        ###   ########.fr       */
+/*   Updated: 2021/07/26 23:02:01 by parkjaekw        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -44,106 +44,6 @@ static void	analyze_quote(t_lexer *lexer, char c, int i)
 		lexer->lex[i] = 'c';
 }
 
-	/*
-	 *2 == hdoc , 3 == append redir
-	 */
-
-int		check_precede_fd(char *cmd, int i)
-{
-	int start;
-
-	start = 0;
-	if (!cmd)
-		return (0);
-	while (i >= 0)
-	{
-		if ((ft_isnum(cmd[i])) == 1)
-		{
-			if (start == 0)
-				start = 1;
-			i--;
-		}
-		else
-			break;
-	}
-	if (start == 1 && i == -1)
-		return (1);
-	else if (start == 1 && i >= 0 && (ft_strchr(" <>", cmd[i])))
-		return (1);
-	return (0);
-}
-
-int		check_follow_fd(char *cmd, int i)
-{
-	int size;
-	int start;
-
-	size = 0;
-	start = 0;
-	if (!cmd || cmd[i] != '&')
-		return (0);
-	size = ft_strlen(cmd);
-	if (i >= size)
-		return (0);
-	printf("ok\n");
-	if (cmd[i] == '&')
-		i++;
-	while (i < size)
-	{
-		if ((ft_isnum(cmd[i])) == 1)
-			start = 1;
-		else
-			break;
-		i++;
-	}
-	if (start == 1 && cmd[i] == '\0')
-		return (1);
-	else if (start == 1 && i < size && ft_strchr(" <>", cmd[i]))
-		return (1);
-	return (0);
-}
-
-void		mark_precede_fd(char *lex, char type, int i)
-{
-	if (!lex)
-		return ;
-	while (i >= 0)
-	{
-		if ((ft_isnum(g_sh.cmd[i])) == 1)
-		{
-			lex[i] = type;
-			i--;
-		}
-		else
-			break;
-	}
-}
-
-void		mark_follow_fd(char *lex, char type, int *i)
-{
-	int size;
-
-	size = ft_strlen(g_sh.cmd);
-	(*i)++;
-	if (*i >= size)
-		return ;
-	if (g_sh.cmd[*i] == '&')
-		lex[*i] = 'f';
-	printf("mark follow fd\n");
-	if (!lex)
-		return ;
-	if (g_sh.cmd[*i] == '&')
-		lex[(*i)++] = '&';
-	while (*i < size)
-	{
-		if ((ft_isnum(g_sh.cmd[*i])) == 1)
-			lex[*i] = type;
-		else
-			break ;
-		(*i)++;
-	}
-}
-
 static void	analyze_operator(char *lex, int ret, int *i)
 {
 	if (ret == 1)
@@ -154,7 +54,7 @@ static void	analyze_operator(char *lex, int ret, int *i)
 		(*i)++;
 		lex[*i] = 'H';
 	}
-	else if(ret == S_AREDIR)
+	else if (ret == S_AREDIR)
 	{
 		lex[*i] = 'A';
 		(*i)++;
@@ -166,7 +66,7 @@ static void	analyze_operator(char *lex, int ret, int *i)
 			lex[*i] = 'I';
 		else if (ret == S_OREDIR)
 			lex[*i] = 'O';
-		}
+	}
 }
 
 void	analyze_command(t_lexer *lexer, char *cmd, int *i)
