+++ srcs/parse/tokenizer.c
@@ -6,7 +6,7 @@
 /*   By: jaekpark <jaekpark@student.42seoul.fr      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/04 18:28:05 by jaekpark          #+#    #+#             */
-/*   Updated: 2021/07/25 13:28:51 by parkjaekw        ###   ########.fr       */
+/*   Updated: 2021/07/26 22:35:26 by parkjaekw        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,35 +61,6 @@ static void		set_type(t_lst *token)
 			tmp->type = 'S';
 		tmp = tmp->next;
 	}
-}
-
-int				analyze_append_redir(t_token *node)
-{
-	int	fd;
-	int	cnt;
-	int	i;
-
-	i = 0;
-	cnt = 0;
-	fd = -2;
-	if (!node)
-		return (-1);
-	else if ((ft_isnum(node->token[0])) == 1)
-		fd = ft_atoi(node->token);
-	if (fd == -1)
-		printf("BraveShell: file descriptor out of range: Bad file "
-				"descriptor\n");
-	else if (fd > 255)
-		printf("BraveShell: %d: Bad file descriptor\n", fd);
-	while ((ft_isnum(node->token[i])) == 1)
-		i++;
-	while (node->token[i])
-	{
-		cnt++;
-		i++;
-	}
-	printf("mark cnt = %d\n", cnt);
-	return (cnt);
 }
 
 int				check_unexpected_token(t_token *node)
@@ -154,15 +125,6 @@ int				analyze_syntax(t_lst *token)
 		tmp = tmp->next;
 	}
 	tmp = token->head;
-	/*while (tmp)*/
-	/*{*/
-		/*if (tmp->type == 'F')*/
-			/*ret = check_fd_token(tmp);*/
-		/*if (ret == -1)*/
-			/*return (-1);*/
-		/*tmp = tmp->next;*/
-	/*}*/
-	/*tmp = token->head;*/
 	while (tmp)
 	{
 		if (tmp->type == 'S')
@@ -246,49 +208,6 @@ char	*unclosed_pipe(void)
 	}
 	return (NULL);
 }
-
-/*
- *int			check_last_token(void)
- *{
- *    char	*line;
- *    t_token	*node;
- *    t_lexer	*tmp;
- *    int		ret;
- *
- *    node = g_sh.token->tail;
- *    if (!node)
- *        return (-1);
- *    else if (node->token)
- *    {
- *        if (node->token[0] == '|')
- *        {
- *            while (TRUE)
- *            {
- *                signal(SIGINT, SIG_DFL);
- *                signal(SIGQUIT, SIG_IGN);
- *                line = readline("> ");
- *                if (ft_strlen(line) > 0)
- *                {
- *                    g_sh.cmd = ft_strjoin_sp(g_sh.cmd, line);
- *                    tmp = lexer(g_sh.cmd);
- *                    free_lexer(g_sh.lexer);
- *                    printf("cmd = %s, tmp lex = %s\n",g_sh.cmd, tmp->lex);
- *                    g_sh.lexer = tmp;
- *                    free_token(g_sh.token);
- *                    g_sh.token = malloc(sizeof(t_lst));
- *                    init_lst(g_sh.token);
- *                    ret = tokenizer(tmp->lex);
- *                    free(line);
- *                    break ;
- *                }
- *                rl_redisplay();
- *                free(line);
- *            }
- *        }
- *    }
- *    return (analyze_syntax(g_sh.token));
- *}
- */
 
 int				tokenizer(char *lex)
 {
