+++ srcs/setup/set.c
@@ -6,13 +6,36 @@
 /*   By: parkjaekwang <marvin@42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/07/13 19:26:58 by parkjaekw         #+#    #+#             */
-/*   Updated: 2021/07/26 20:40:03 by parkjaekw        ###   ########.fr       */
+/*   Updated: 2021/07/26 22:20:44 by parkjaekw        ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-extern t_conf g_sh;
+extern t_conf	g_sh;
+
+static int	set_cmd_after_pipe(void)
+{
+	char	*new_cmd;
+	int		ret;
+
+	ret = 0;
+	new_cmd = NULL;
+	free_lexer(g_sh.lexer);
+	g_sh.lexer = NULL;
+	free_token(g_sh.token);
+	g_sh.token = NULL;
+	if (g_sh.token == NULL)
+	{
+		g_sh.token = malloc(sizeof(t_lst));
+		init_lst(g_sh.token);
+	}
+	new_cmd = unclosed_pipe();
+	g_sh.cmd = ft_strjoin_sp(g_sh.cmd, new_cmd);
+	g_sh.lexer = lexer(g_sh.cmd);
+	ret = tokenizer(g_sh.lexer->lex);
+	return (ret);
+}
 
 void	set_prompt(void)
 {
@@ -26,8 +49,7 @@ void	set_prompt(void)
 void	set_process(void)
 {
 	t_token	*tmp;
-	int	ret;
-	char *new_cmd;
+	int		ret;
 
 	ret = 0;
 	tmp = NULL;
@@ -35,22 +57,8 @@ void	set_process(void)
 	if (g_sh.lexer->err == 1)
 		exit_shell(0);
 	ret = tokenizer(g_sh.lexer->lex);
-	if (ret == 2)
-	{
-		free_lexer(g_sh.lexer);
-		g_sh.lexer = NULL;
-		free_token(g_sh.token);
-		g_sh.token = NULL;
-		if (g_sh.token == NULL)
-		{
-			g_sh.token = malloc(sizeof(t_lst));
-			init_lst(g_sh.token);
-		}
-		new_cmd = unclosed_pipe();
-		g_sh.cmd = ft_strjoin_sp(g_sh.cmd, new_cmd);
-		g_sh.lexer = lexer(g_sh.cmd);
-		ret = tokenizer(g_sh.lexer->lex);
-	}
+	while (ret == 2)
+		ret = set_cmd_after_pipe();
 	add_history(g_sh.cmd);
 	rl_redisplay();
 	if (ret == -1)
@@ -75,7 +83,7 @@ void	set_env(char **envp)
 	g_sh.envp = envp;
 	while (envp[++i])
 	{
-		tmp = ft_split_env(envp[i]);
+		tmp = split_env(envp[i]);
 		make_env(env, tmp[0], tmp[1]);
 		ft_free_double((void **)tmp);
 	}
