+++ srcs/temp.c
@@ -120,3 +120,174 @@
 	/*save_process(process, syn);*/
 	/*return (tmp);*/
 /*}*/
+
+
+
+int	check_precede_fd(char *cmd, int i)
+{
+	int	start;
+
+	start = 0;
+	if (!cmd)
+		return (0);
+	while (i >= 0)
+	{
+		if ((ft_isnum(cmd[i])) == 1)
+		{
+			if (start == 0)
+				start = 1;
+			i--;
+		}
+		else
+			break ;
+	}
+	if (start == 1 && i == -1)
+		return (1);
+	else if (start == 1 && i >= 0 && (ft_strchr(" <>", cmd[i])))
+		return (1);
+	return (0);
+}
+
+int	check_follow_fd(char *cmd, int i)
+{
+	int	size;
+	int	start;
+
+	size = 0;
+	start = 0;
+	if (!cmd || cmd[i] != '&')
+		return (0);
+	size = ft_strlen(cmd);
+	if (i >= size)
+		return (0);
+	printf("ok\n");
+	if (cmd[i] == '&')
+		i++;
+	while (i < size)
+	{
+		if ((ft_isnum(cmd[i])) == 1)
+			start = 1;
+		else
+			break ;
+		i++;
+	}
+	if (start == 1 && cmd[i] == '\0')
+		return (1);
+	else if (start == 1 && i < size && ft_strchr(" <>", cmd[i]))
+		return (1);
+	return (0);
+}
+
+void		mark_precede_fd(char *lex, char type, int i)
+{
+	if (!lex)
+		return ;
+	while (i >= 0)
+	{
+		if ((ft_isnum(g_sh.cmd[i])) == 1)
+		{
+			lex[i] = type;
+			i--;
+		}
+		else
+			break;
+	}
+}
+
+void		mark_follow_fd(char *lex, char type, int *i)
+{
+	int size;
+
+	size = ft_strlen(g_sh.cmd);
+	(*i)++;
+	if (*i >= size)
+		return ;
+	if (g_sh.cmd[*i] == '&')
+		lex[*i] = 'f';
+	printf("mark follow fd\n");
+	if (!lex)
+		return ;
+	if (g_sh.cmd[*i] == '&')
+		lex[(*i)++] = '&';
+	while (*i < size)
+	{
+		if ((ft_isnum(g_sh.cmd[*i])) == 1)
+			lex[*i] = type;
+		else
+			break ;
+		(*i)++;
+	}
+}
+
+int				analyze_append_redir(t_token *node)
+{
+	int	fd;
+	int	cnt;
+	int	i;
+
+	i = 0;
+	cnt = 0;
+	fd = -2;
+	if (!node)
+		return (-1);
+	else if ((ft_isnum(node->token[0])) == 1)
+		fd = ft_atoi(node->token);
+	if (fd == -1)
+		printf("BraveShell: file descriptor out of range: Bad file "
+				"descriptor\n");
+	else if (fd > 255)
+		printf("BraveShell: %d: Bad file descriptor\n", fd);
+	while ((ft_isnum(node->token[i])) == 1)
+		i++;
+	while (node->token[i])
+	{
+		cnt++;
+		i++;
+	}
+	printf("mark cnt = %d\n", cnt);
+	return (cnt);
+}
+
+/*
+ *int			check_last_token(void)
+ *{
+ *    char	*line;
+ *    t_token	*node;
+ *    t_lexer	*tmp;
+ *    int		ret;
+ *
+ *    node = g_sh.token->tail;
+ *    if (!node)
+ *        return (-1);
+ *    else if (node->token)
+ *    {
+ *        if (node->token[0] == '|')
+ *        {
+ *            while (TRUE)
+ *            {
+ *                signal(SIGINT, SIG_DFL);
+ *                signal(SIGQUIT, SIG_IGN);
+ *                line = readline("> ");
+ *                if (ft_strlen(line) > 0)
+ *                {
+ *                    g_sh.cmd = ft_strjoin_sp(g_sh.cmd, line);
+ *                    tmp = lexer(g_sh.cmd);
+ *                    free_lexer(g_sh.lexer);
+ *                    printf("cmd = %s, tmp lex = %s\n",g_sh.cmd, tmp->lex);
+ *                    g_sh.lexer = tmp;
+ *                    free_token(g_sh.token);
+ *                    g_sh.token = malloc(sizeof(t_lst));
+ *                    init_lst(g_sh.token);
+ *                    ret = tokenizer(tmp->lex);
+ *                    free(line);
+ *                    break ;
+ *                }
+ *                rl_redisplay();
+ *                free(line);
+ *            }
+ *        }
+ *    }
+ *    return (analyze_syntax(g_sh.token));
+ *}
+ */
+
